fixed region in apply

+++ tangle_cnn/wildcat_main.py
@@ -192,8 +192,8 @@ class TrainedWildcat:
             region = list(float(val) for val in region)
             if all(val <= 1.0 for val in region):
                 u_range, v_range = (
-                    (slide_dim[i] * region[i] // window_size_raw,
-                     1 + slide_dim[i] * (region[i] + region[i+2]) // window_size_raw) for i in range(2))
+                    (int(slide_dim[i] * region[i] // window_size_raw),
+                     int(1 + slide_dim[i] * (region[i] + region[i+2]) // window_size_raw)) for i in range(2))
             else:
                 u_range = (int(region[0]), int(region[0] + region[2]))
                 v_range = (int(region[1]), int(region[1] + region[3]))
